# apps/api/src/swagger/api-schema.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "description": "Returns all user accounts. Requires permission: users:read",
        "operationId": "UsersController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List users",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{uuid}": {
      "get": {
        "description": "Returns a user by UUID. Requires permission: users:read",
        "operationId": "UsersController_getUser",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{uuid}/role": {
      "patch": {
        "description": "Replaces the user role. A user can have only one role. Requires permission: users.roles:update",
        "operationId": "UsersController_updateUserRole",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "User or role not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user role",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Creates a user account. The first registered account receives the ADMIN role; later users receive USER.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "summary": "Register user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates by email or username and returns a JWT access token.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Log in",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Revokes the current JWT by adding its token id to the Redis blacklist.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log out",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/projects": {
      "get": {
        "description": "Fetches and returns an array of all project entities stored in the system.",
        "operationId": "ProjectsController_getProjects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all projects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetProjectResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "summary": "Retrieve a list of all projects",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "description": "Creates a new project entity with the provided data. Requires permission: projects:create",
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{uuid}": {
      "get": {
        "description": "Fetches and returns a single project entity based on the provided ID.",
        "operationId": "ProjectsController_getProjectById",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project with the specified ID not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "summary": "Retrieve a project",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "description": "Updates the project entity identified by the provided ID with new data.  Requires permission: projects:update",
        "operationId": "ProjectsController_updateProject",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProjectResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Project with the specified ID not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an existing project",
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "description": "Deletes the project entity identified by the provided ID. Requires permission: projects:delete",
        "operationId": "ProjectsController_deleteProject",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden, insufficient permissions"
          },
          "404": {
            "description": "Project with the specified ID not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service is unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a project",
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio website API",
    "description": "API documentation for the portfolio website",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "UserRoleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ADMIN"
          },
          "label": {
            "type": "string",
            "example": "Administrator"
          },
          "permissions": {
            "example": [
              "projects:create",
              "projects:update"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "label",
          "permissions"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "example": "e2718142-3c3f-4a5f-9038-f53ea9f15b47"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "isActive": {
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/UserRoleResponse"
          }
        },
        "required": [
          "uuid",
          "email",
          "username",
          "isActive",
          "role"
        ]
      },
      "UpdateUserRoleBody": {
        "type": "object",
        "properties": {
          "roleCode": {
            "type": "string",
            "example": "ADMIN"
          }
        },
        "required": [
          "roleCode"
        ]
      },
      "RegisterBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "VeryStrongPassword123!"
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "user"
        ]
      },
      "LoginBody": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "example": "VeryStrongPassword123!"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "revoked": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "revoked"
        ]
      },
...
      "GetProjectResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "description": "Unique identifier for the project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "uuid"
        ]
      },
      "CreateProjectBody": {
        "type": "object",
        "properties": {}
      },
      "UpdateProjectBody": {
        "type": "object",
        "properties": {}
      },
      "UpdateProjectResponse": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}